add input validation and command safety checks

+++ src/TheMain.java
@@ -4,39 +4,51 @@ import java.util.Scanner;
 
 
 public class TheMain {
-    public static void main(String[] args) throws IOException {
+    public static void main(String[] args) throws IOException{
         Persistence p = new  Persistence();
         Database db = new Database();
         try{
             p.readFromFile(db.getStorage());
-            System.out.println(db.GET("name"));
-            System.out.println(db.KEYS());
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
         Scanner sc = new  Scanner(System.in);
-//        Database db = new Database();
 
         boolean flag = true;
         while(flag) {
             String input = sc.nextLine();
+            if(input.trim().isEmpty()) {
+                continue;
+            }
             List<String> words = List.of(input.split("\\s+"));
 
             switch (words.get(0)) {
                 case "SET": {
-                    db.SET(words.get(1), words.get(2));
+                    if(words.size() < 3)
+                        System.out.println("Error: SET requires a Key and a Value");
+                    else
+                        db.SET(words.get(1), words.get(2));
                     break;
                 }
                 case "GET": {
-                    System.out.println(db.GET(words.get(1)));
+                    if(words.size()<2)
+                        System.out.println("Error: GET requires a KEY");
+                    else
+                        System.out.println(db.GET(words.get(1)));
                     break;
                 }
                 case "DELETE": {
-                    db.DELETE(words.get(1));
+                    if(words.size()<2)
+                        System.out.println("Error: DELETE requires a KEY");
+                    else
+                        db.DELETE(words.get(1));
                     break;
                 }
                 case "EXISTS": {
-                    System.out.println(db.EXISTS(words.get(1)));
+                    if(words.size()<2)
+                        System.out.println("Error: EXISTS requires a KEY");
+                    else
+                        System.out.println(db.EXISTS(words.get(1)));
                     break;
                 }
                 case "EXIT":{
